riski-backend/pyproject.toml
[project]
name = "riski-backend"
license = "MIT"
version = "0.1.0"
description = "Python backend for finding information in munich's political information system RIS"
readme = "README.md"
requires-python = "==3.13.9"
authors = [
    { name = "Fabian Reinold", email = "fabian.reinold@muenchen.de" },
    { name = "Stefan Hentschel", email = "stefan.hentschel@muenchen.de" },
    { name = "Annalena Wittmann", email = "annalena.wittmann@muenchen.de" },
    { name = "Tim-Luca Pfeifer", email = "tim-luca.pfeifer@muenchen.de" },
    { name = "Michael Jaumann", email = "michael.jaumann@muenchen.de" },
]
maintainers = [
   { name = "Stefan Hentschel", email = "stefan.hentschel@muenchen.de" },
]

dependencies = [
    "fastapi==0.120.2",
    "uvicorn==0.38.0",
    "httpx==0.28.1",
    "cryptography==46.0.3",
    "pydantic-settings==2.12.0",
    "pydantic==2.12.4",
    "python-dotenv==1.2.1",
    "truststore==0.10.4",
]

[dependency-groups]
dev = [
    "pre-commit==4.5.0",
    "pytest-asyncio==1.3.0",
    "pytest==9.0.1",
    "ruff==0.14.6",
]

[tool.uv]
add-bounds = "exact"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]
asyncio_mode = "auto"